Add xdg-compliance, bash_profile, and bashrc
xdg-compliance is sourced to set the XDG variables. bash_profile uses
them to set the location of the Xorg logs. /etc/profile should be
modified to source ~/.config/bash/bash_profile to bootstrap the process.
If this is not possible, ~/.bash_profile can be created as a symlink
pointing to $XDG_CONFIG_HOME/bash/bash_profile. bashrc sets environment
variables and creates aliases to make various programs store their files
in the appropriate places. A script symlinked to ~/bin/bash ensures that
bash always reads the bashrc file from the appropriate XDG location.